+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- etc-hosts: add a subcomand to generate /etc/hosts entries for the public bridge or a private one
 
 ## [0.4.0] 2019-10-11
 ### Added